Reorganize webpack resources per layout
