+++ app/webpacker/src/javascript/stimulus/hello_controller.js
