+++ app/webpacker/src/javascript/stimulus/list_items_controller.js
